Update cisa_cve_compact_json.py
Script downloads the latest catalog of known exploited CVE's tracked by CISA.  If not file exists, it will write the latest to disk.  Otherwise, it will check to see if a newer catalog version than what is on disk, exists.  If so, it will perform the update, if not, it will keep it's current version.